backend/services/user-service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
python-multipart
aiosmtplib
jinja2
argon2-cffi
PyJWT
email-validator
redis==5.0.1
hiredis==2.3.2
tenacity==8.2.3
bleach==6.1.0
python-magic==0.4.27
email-validator==2.1.0
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-gcp-trace==1.6.0
cryptography==41.0.7
secure==0.3.0
